+++ R/utils.R
@@ -40,17 +40,33 @@ modify_list <- function(.x, ...) {
 }
 
 
-sys_sleep <- function(seconds) {
+sys_sleep <- function(seconds, fps = 10) {
   check_number(seconds, "`seconds`")
 
-  if (seconds > 0) {
-    # TODO: add progress bar
-    signal("", class = "httr2_sleep", seconds = seconds)
-    Sys.sleep(seconds)
+  if (seconds == 0) {
+    return(invisible())
   }
+
+  signal("", class = "httr2_sleep", seconds = seconds)
+
+  cli::cli_progress_bar(
+    format = "Waiting {seconds}s to retry {cli::pb_bar}",
+    total = seconds * fps
+  )
+
+  start <- cur_time()
+  until <- start + seconds
+  while(cur_time() < until) {
+    Sys.sleep(1 / fps)
+    elapsed <- cur_time() - start
+    cli::cli_progress_update(set = elapsed * fps)
+  }
+  cli::cli_progress_done()
 
   invisible()
 }
+
+cur_time <- function() proc.time()[[3]]
 
 check_string <- function(x, name, optional = TRUE) {
   if (is_string(x) && !is.na(x)) {